requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
beautifulsoup4>=4.12.0
requests>=2.31.0
PyYAML>=6.0.1
waveshare-epd>=1.0.0
lxml>=4.9.0
